requirements-whatsapp.txt
# WhatsApp Integration Requirements
# Additional dependencies for WhatsApp Business API integration

# Core web framework for webhook server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# HTTP client for WhatsApp API calls
aiohttp>=3.9.0
requests>=2.31.0

# Data validation and serialization
pydantic>=2.5.0

# File handling for session storage
aiofiles>=23.2.0

# Environment variables
python-dotenv>=1.0.0

# Optional: Database support (uncomment if needed)
# asyncpg>=0.29.0  # PostgreSQL
# psycopg2-binary>=2.9.9  # PostgreSQL (sync)
# pymongo>=4.6.0  # MongoDB
# motor>=3.3.2  # MongoDB (async)

# Optional: Redis support for session caching (uncomment if needed)
# redis>=5.0.1
# aioredis>=2.0.1

# Optional: Enhanced logging and monitoring (uncomment if needed)
# structlog>=23.2.0
# prometheus-client>=0.19.0

# Security and rate limiting (optional)
# slowapi>=0.1.9  # Rate limiting
# cryptography>=41.0.8  # Enhanced security

# Development and testing dependencies
# pytest>=7.4.3
# pytest-asyncio>=0.21.1
# httpx>=0.25.2  # For testing FastAPI
# pytest-mock>=3.12.0

